requirements.txt
Django
djangorestframework
django-rest-swagger
django-filter
markdown
django-cors-headers
django-versatileimagefield
django-model-utils
djoser
requests
Pillow
python-dotenv
drf-writable-nested
django-constant[database]
psycopg2-binary
django-ckeditor